chatbots/pyproject.toml
[project]
name = "gpt"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Dan Drennan", email = "danjdrennan@gmail.com"},
]
dependencies = [
    "openai>=1.40.6",
    "python-dotenv>=1.0.1",
    "anthropic>=0.34.1",
    "ipython>=8.26.0",
]
readme = "README.md"
license = {text = "MIT"}
requires-python = "==3.12.*"

[tool.pdm]
distribution = false

[tool.basedpyright]
typeCheckingMode = "basic"
